# Conflicts:
#	WalletKitCore/src/bitcoin/BRSyncManager.c
#	WalletKitCore/src/bitcoin/BRWalletManager.c
#	WalletKitCore/src/bitcoin/BRWalletManager.h
#	WalletKitCore/src/crypto/BRCryptoAddress.c
#	WalletKitCore/src/crypto/BRCryptoHash.c
#	WalletKitCore/src/crypto/BRCryptoNetwork.c
#	WalletKitCore/src/crypto/BRCryptoNetworkP.h
#	WalletKitCore/src/crypto/BRCryptoPayment.c
#	WalletKitCore/src/crypto/BRCryptoWallet.c
#	WalletKitCore/src/crypto/BRCryptoWalletManagerClient.c
#	WalletKitCore/src/crypto/handlers/eth/BRCryptoWalletManagerPersistETH.c
#	WalletKitCore/src/ethereum/ewm/BREthereumClient.h
#	WalletKitCore/src/ethereum/ewm/BREthereumEWM.c
#	WalletKitCore/src/ethereum/ewm/BREthereumEWM.h
#	WalletKitCore/src/ethereum/ewm/BREthereumEWMClient.c
#	WalletKitCore/src/ethereum/ewm/BREthereumEWMEvent.c
#	WalletKitCore/src/ethereum/ewm/BREthereumEWMPrivate.h
#	WalletKitCore/src/ethereum/ewm/BREthereumTransfer.c
#	WalletKitCore/src/ethereum/ewm/BREthereumTransfer.h
#	WalletKitCore/src/ethereum/ewm/BREthereumWallet.c
#	WalletKitCore/src/ethereum/ewm/BREthereumWallet.h
#	WalletKitCore/src/generic/BRGenericBase.h
#	WalletKitCore/src/generic/BRGenericHedera.c
#	WalletKitCore/src/generic/BRGenericManager.c
#	WalletKitCore/src/generic/BRGenericRipple.c
#	WalletKitJava/corecrypto/src/main/java/com/breadwallet/corecrypto/System.java
#	WalletKitSwift/WalletKit/BRCryptoSystem.swift

+++ WalletKitCore/src/ethereum/les/BREthereumLES.c
@@ -597,7 +597,7 @@ lesCreate (BREthereumNetwork network,
     // preserved.
     nodeEndpointSetStatus (les->localEndpoint,
                            messageP2PStatusCreate (0x00,  // ignored
-                                                   ethNetworkGetChainId(network),
+                                                   (uint64_t) ethNetworkGetChainId(network),
                                                    les->head.number,
                                                    les->head.hash,
                                                    les->head.totalDifficulty,
